+++ oppijanumerorekisteri-service/src/main/java/fi/vm/sade/oppijanumerorekisteri/services/OppijanumerorekisteriBusinessService.java
@@ -1,12 +1,4 @@
 package fi.vm.sade.oppijanumerorekisteri.services;
 
-import DTOs.HenkiloPerustietoDto;
-
-import java.util.List;
-
 public interface OppijanumerorekisteriBusinessService {
-    Boolean getHasHetu(String oid);
-    boolean getOidExists(String oid);
-    String getOidByHetu(String hetu);
-    List<HenkiloPerustietoDto> getHenkiloPerustietosByOids(List<String> oids);
 }
